pyproject.toml
[tool.poetry]
authors = ["safeluren"]
name = "nonebot-plugin-bf1marneserverlist"
version = "0.1.23"
description = "nonebot-plugin-bf1marneserverlist"
readme = "README.md"
packages = [{include = "nonebot_plugin_bf1marneserverlist"}]

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "^2.2.0"
nonebot-adapter-onebot = "^2.4.0"
httpx = "^0.27.0"
nonebot_plugin_localstore = "^0.6.0"


[tool.nonebot]
plugins = ["nonebot_plugin_localstore"]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
plugin_dirs = ["nonebot_plugin_bf1marneserverlist"]
builtin_plugins = []

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
